dropped template library for aduser/service accounts
this is a compliation of scripts for powershell to do ad functions. as a template for starting development for api calls to do ad functions from a secure location. no data for my company is included, as these were written from scratch. this is just a template starting point for the necessary pieces to do each function.